medical-report-agent/requirements.txt
anthropic==0.40.0
python-docx==1.1.0
matplotlib==3.8.2
pandas==2.1.4
reportlab==4.0.9
Pillow==10.1.0
numpy==1.26.2
